# docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResult:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      fingerprint:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - fingerprint
    - password
    - username
    type: object
  dto.LogoutRequest:
    properties:
      fingerprint:
        type: string
      refresh_token:
        type: string
    required:
    - fingerprint
    - refresh_token
    type: object
  dto.MonitorRequest:
    properties:
      expected_response: {}
      interval:
        minimum: 0
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      request_spec: {}
      target:
        type: string
      timeout:
        minimum: 0
        type: integer
      type:
        type: string
    required:
    - interval
    - name
    - request_spec
    - target
    - timeout
    - type
    type: object
  dto.MonitorResponse:
    properties:
      id:
        type: integer
    type: object
  dto.RefreshRequest:
    properties:
      fingerprint:
        type: string
      refresh_token:
        type: string
    required:
    - fingerprint
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      fingerprint:
        type: string
      password:
        minLength: 6
        type: string
      telegram_id:
        type: integer
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - fingerprint
    - password
    - username
    type: object
  models.Monitor:
    properties:
      id:
        type: integer
      interval:
        type: integer
      is_active:
        type: boolean
      last_checked_at:
        type: string
      name:
        type: string
      target:
        type: string
      timeout:
        type: integer
      type:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for uptime monitoring service
  title: Uptime Monitoring API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: user credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: tokens
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: tokens
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: user credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register user
      tags:
      - auth
  /monitors:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Monitor'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all user's monitors
      tags:
      - monitors
    post:
      consumes:
      - application/json
      parameters:
      - description: monitor create request
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/dto.MonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MonitorResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new monitor
      tags:
      - monitors
  /monitors/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete monitor
      tags:
      - monitors
    get:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Monitor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get monitor by ID
      tags:
      - monitors
    put:
      consumes:
      - application/json
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      - description: monitor update request
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/dto.MonitorRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update monitor
      tags:
      - monitors
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите токен в формате: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"